It's a mess but it's my mess

+++ bash/purplpkg.sh
@@ -149,16 +149,51 @@ if grep -q "<head><title>404 Not Found</title></head>" "$2".ppkg; then
  rm versions/"$2"
  rm files/"$2"
  echo "Check the package name and try again."
- export 2_NOT_FOUND=1
+ export TWO_NOT_FOUND=1
 fi
 
-if grep -q "<head><title>404 Not Found</title></head>" "$3".ppkg; then
- echo "Package "$3" not found on server."
- rm "$3".ppkg
- rm versions/"$3"
- rm files/"$3"
- echo "Check the package name and try again."
- export 3_NOT_FOUND=1
+if [ -f "/data/purplpkg/"$3".ppkg" ]; then
+ if grep -q "<head><title>404 Not Found</title></head>" "$3".ppkg; then
+  echo "Package "$3" not found on server."
+  rm "$3".ppkg
+  rm versions/"$3"
+  rm files/"$3"
+  echo "Check the package name and try again."
+  export THREE_NOT_FOUND=1
+ fi
 fi
 
+if [ "$THREE_NOT_FOUND" == "1" ] && [ "$TWO_NOT_FOUND" == "1" ]; then
+ exit 1
+fi
 
+if [ "$THREE_NOT_FOUND" == "1" ]; then
+    echo "Installing..."
+    tar -xzf "$2".ppkg
+    echo "Cleaning up..."
+    rm "$2".ppkg
+    echo "Package "$2" installed with version "$VERSION""
+    #Lower frequency back to "balanced" wire_d preset
+    echo 533333 > /sys/devices/system/cpu/cpu0/cpufreq/scaling_max_freq
+    exit 0
+elif [ "$TWO_NOT_FOUND" == "1" ]; then
+    echo "Installing..."
+    tar -xzf "$3".ppkg
+    echo "Cleaning up..."
+    rm "$3".ppkg
+    echo "Package "$3" installed"
+    #Lower frequency back to "balanced" wire_d preset
+    echo 533333 > /sys/devices/system/cpu/cpu0/cpufreq/scaling_max_freq
+    exit 0
+elif [ ! "$THREE_NOT_FOUND" == "1" ] && [ ! "$TWO_NOT_FOUND" == "1" ]; then
+    echo "Installing "$2" and "$3""
+    tar -xzf "$2".ppkg
+    tar -xzf "$3".ppkg
+    echo "Cleaning up..."
+    rm "$2".ppkg
+    rm "$3".ppkg
+    echo "Packages "$2" and "$3" installed"
+    #Lower frequency back to "balanced" wire_d preset
+    echo 533333 > /sys/devices/system/cpu/cpu0/cpufreq/scaling_max_freq
+    exit 0
+fi
